Updated jquery.prettyLoader.js
Error - cur_x is not defined

Cause - When the mouse is positioned at a place not belonging to the clientarea. The event parameter does not contain clientX and clientY properties, both are not defined and thus the PrettyLoader plugins throws an error at line#36 and line#37.

Solution - TO correct this problem, I have added a simple condition that checks for such scenario.